requirements.txt
"mcp[cli]"
requests
openai
python-dotenv